+++ src/vehicles/MilitaryAirplane.java
@@ -9,7 +9,7 @@ import enums.Weapons;
 public class MilitaryAirplane extends Airplane {
     private Weapons weapons;
 
-    public MilitaryAirplane(int x, int y, int id, int amountOfStaff, double currentFuel, double maxFuel, Airport destination, Weapons weapons, double speed) {
+    public MilitaryAirplane(double x, double y, int id, int amountOfStaff, double currentFuel, double maxFuel, Airport destination, Weapons weapons, double speed) {
         super(x, y, id, amountOfStaff, currentFuel, maxFuel, destination, speed);
         this.weapons = weapons;
     }
